refactor: separate progressbar
